pyproject.toml
[project]
name = "crmlite"
version = "0.1.0"
description = ""
authors = [
    {name = "SagebrushOcean",email = "katya.shevchenko.2014@inbox.ru"}
]
requires-python = ">=3.12"
dependencies = [
    "django (>=6.0.2,<7.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
